+++ internal/batches/service/service.go
@@ -466,46 +466,33 @@ func (e *duplicateBranchesErr) Error() string {
 	return out.String()
 }
 
-func (svc *Service) ParseBatchSpec(dir string, data []byte, isRemote bool) (*batcheslib.BatchSpec, error) {
+func (svc *Service) ParseBatchSpec(dir string, data []byte) (*batcheslib.BatchSpec, error) {
 	spec, err := batcheslib.ParseBatchSpec(data)
 	if err != nil {
 		return nil, errors.Wrap(err, "parsing batch spec")
 	}
-	if err = validateMount(dir, spec, isRemote); err != nil {
+	if err = validateMount(dir, spec); err != nil {
 		return nil, errors.Wrap(err, "handling mount")
 	}
 	return spec, nil
 }
 
-func validateMount(batchSpecDir string, spec *batcheslib.BatchSpec, isRemote bool) error {
+func validateMount(batchSpecDir string, spec *batcheslib.BatchSpec) error {
 	for i, step := range spec.Steps {
-		for j, mount := range step.Mount {
-			if isRemote {
-				return errors.New("mounts are not supported for server-side processing")
-			}
-			p := mount.Path
-			if !filepath.IsAbs(p) {
+		for _, mount := range step.Mount {
+			if !filepath.IsAbs(mount.Path) {
 				// Try to build the absolute path since Docker will only mount absolute paths
-				p = filepath.Join(batchSpecDir, p)
+				mount.Path = filepath.Join(batchSpecDir, mount.Path)
 			}
-			pathInfo, err := os.Stat(p)
+			_, err := os.Stat(mount.Path)
 			if os.IsNotExist(err) {
-				return errors.Newf("step %d mount path %s does not exist", i+1, p)
+				return errors.Newf("step %d mount path %s does not exist", i+1, mount.Path)
 			} else if err != nil {
 				return errors.Wrapf(err, "step %d mount path validation", i+1)
 			}
-			if !strings.HasPrefix(p, batchSpecDir) {
+			if !strings.HasPrefix(mount.Path, batchSpecDir) {
 				return errors.Newf("step %d mount path is not in the same directory or subdirectory as the batch spec", i+1)
 			}
-			// Mounting a directory on Docker must end with the separator. So, append the file separator to make
-			// users' lives easier.
-			if pathInfo.IsDir() && !strings.HasSuffix(p, string(filepath.Separator)) {
-				p += string(filepath.Separator)
-			}
-			// Update the mount path to the absolute path so building the absolute path (above) does not need to be
-			// redone when adding the mount argument to the Docker container.
-			// TODO: Can this mess with caching? We wouldn't be doing that server-side.
-			step.Mount[j].Path = p
 		}
 	}
 	return nil