+++ internal/batches/service/service_test.go
@@ -234,7 +234,6 @@ func TestService_ParseBatchSpec(t *testing.T) {
 		name         string
 		batchSpecDir string
 		rawSpec      string
-		isRemote     bool
 		expectedSpec *batcheslib.BatchSpec
 		expectedErr  error
 	}{
@@ -327,7 +326,7 @@ changesetTemplate:
 						Container: "alpine:3",
 						Mount: []batcheslib.Mount{
 							{
-								Path:       filepath.Join(tempDir, "sample.sh"),
+								Path:       "./sample.sh",
 								Mountpoint: "/tmp/sample.sh",
 							},
 						},
@@ -371,7 +370,7 @@ changesetTemplate:
 						Container: "alpine:3",
 						Mount: []batcheslib.Mount{
 							{
-								Path:       tempDir + string(filepath.Separator),
+								Path:       tempDir,
 								Mountpoint: "/tmp",
 							},
 						},
@@ -415,7 +414,7 @@ changesetTemplate:
 						Container: "alpine:3",
 						Mount: []batcheslib.Mount{
 							{
-								Path:       tempDir + string(filepath.Separator),
+								Path:       "./",
 								Mountpoint: "/tmp",
 							},
 						},
@@ -461,11 +460,11 @@ changesetTemplate:
 						Container: "alpine:3",
 						Mount: []batcheslib.Mount{
 							{
-								Path:       filepath.Join(tempDir, "sample.sh"),
+								Path:       "./sample.sh",
 								Mountpoint: "/tmp/sample.sh",
 							},
 							{
-								Path:       filepath.Join(tempDir, "another.sh"),
+								Path:       "./another.sh",
 								Mountpoint: "/tmp/another.sh",
 							},
 						},
@@ -523,32 +522,10 @@ changesetTemplate:
 `, tempOutsideDir),
 			expectedErr: errors.New("handling mount: step 1 mount path is not in the same directory or subdirectory as the batch spec"),
 		},
-		{
-			name:         "mount remote processing",
-			batchSpecDir: tempDir,
-			rawSpec: `
-name: test-spec
-description: A test spec
-steps:
-  - run: /tmp/foo/bar/sample.sh
-    container: alpine:3
-    mount:
-      - path: /valid/sample.sh
-        mountpoint: /tmp/foo/bar/sample.sh
-changesetTemplate:
-  title: Test Mount
-  body: Test a mounted path
-  branch: test
-  commit:
-    message: Test
-`,
-			isRemote:    true,
-			expectedErr: errors.New("handling mount: mounts are not supported for server-side processing"),
-		},
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			spec, err := svc.ParseBatchSpec(test.batchSpecDir, []byte(test.rawSpec), test.isRemote)
+			spec, err := svc.ParseBatchSpec(test.batchSpecDir, []byte(test.rawSpec))
 			if test.expectedErr != nil {
 				assert.Equal(t, test.expectedErr.Error(), err.Error())
 			} else {